arrow: revert sage binaries to default refsw ones.
with the new sage apps/bootloaders for 15.4, we can no
longer use the 'avko' overwrite one for the time being
or we get failure to boot due to app/boot versioning
mismatch.  so we use the default ones from refsw
for the time being until we get the avko special one for
the same versions.

refs #SWSTB-422

Change-Id: I8e9aeae1c866987dde71566114e7fe7f1a9ba7e6
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>