recovery: force graceful startup.
- allow nexus loading and nxmini to boot up slightly
  earlier than the recovery itself (we cannot control
  it finely otherwise).
- wait for nexus to be loaded before allowing recovery
  to go on.

those changes allow to sync the recovery boot up to the
current nexus/platform startup in order to ensure the application
itself does not 'restart' (due to early crash); speeding up the
time to recovery but most importantly sync'ing the recovery boot.

this in particular allows the multi-stage ota sideload to work
properly without skipped steps.

refs #SWANDROID-3351

Change-Id: I9b9d2f54408774810ff393f717d724775f2a9e25
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>