banff: increase the lmk threshold.
by increasing the lmk threshold, we want to force memory reclaim
to kick in earlier which helps with scenarios where a lots of applications
are open one after the other continuously leading to device becoming
sluggish as reclaim/swap takes place.  by increasing the limit of the lmk
we allow more background killing to improve the reclaim/swap time and
need; overall giving a better user experience.

b/29521215
refs #SWANDROID-3449

Change-Id: Icc4c6e1aefc2c658844d726480978ee67356f362
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>