device: fix up 7251 memory map.
on some devices, namely on the single memory controller one's, it is
best to pack the nexus heaps towards the end of the memory bank to
avoid struggling with the nexus heap placement algorithm limitations.

also updated alignment for heaps which is now 1MB when using SAGE 3.x,
as it is the case here.

refs #SWANDROID-3445

Change-Id: I5f9e9d90ae95d125a841116d42abc61798babcd0
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>